Fixed quit and quit dialogue. It's no longer stuck in a dialogue loop.
